requirements.txt
selenium==4.19.0
pytest==8.1.1
python-dotenv==1.0.1